NONE: Revert unnecessary changes
