+++ src/infrastructure/jira/errors.ts
@@ -1,0 +1,54 @@
+import { OperationError } from '../../common/errors';
+import type { AtlassianAssociation } from '../../domain/entities';
+
+export class SubmitDesignJiraOperationError extends OperationError {
+	designId?: string;
+	rejectionErrors?: { readonly message: string }[];
+	unknownIssueKeys?: string[];
+	unknownAssociations?: AtlassianAssociation[];
+
+	private constructor({
+		message,
+		designId,
+		rejectionErrors,
+		unknownIssueKeys,
+		unknownAssociations,
+	}: {
+		message: string;
+		designId?: string;
+		rejectionErrors?: { readonly message: string }[];
+		unknownIssueKeys?: string[];
+		unknownAssociations?: AtlassianAssociation[];
+	}) {
+		super(message);
+		this.designId = designId;
+		this.rejectionErrors = rejectionErrors;
+		this.unknownIssueKeys = unknownIssueKeys;
+		this.unknownAssociations = unknownAssociations;
+	}
+
+	static designRejected(
+		designId: string,
+		rejectionErrors: { readonly message: string }[],
+	): SubmitDesignJiraOperationError {
+		return new SubmitDesignJiraOperationError({
+			message: 'The design submission has been rejected',
+			designId,
+			rejectionErrors,
+		});
+	}
+
+	static unknownIssueKeys(unknownIssueKeys: string[]) {
+		return new SubmitDesignJiraOperationError({
+			message: 'The design has unknown issue keys',
+			unknownIssueKeys,
+		});
+	}
+
+	static unknownAssociations(unknownAssociations: AtlassianAssociation[]) {
+		return new SubmitDesignJiraOperationError({
+			message: 'The design has unknown associations',
+			unknownAssociations,
+		});
+	}
+}
