+++ src/infrastructure/jira/index.ts
@@ -1,1 +1,2 @@
 export * from './jira-service';
+export * from './errors';
