+++ src/infrastructure/jira/jira-service.ts
@@ -1,3 +1,4 @@
+import { SubmitDesignJiraOperationError } from './errors';
 import type {
 	GetIssuePropertyResponse,
 	SubmitDesignsResponse,
@@ -5,7 +6,7 @@ import type {
 import { jiraClient } from './jira-client';
 import { ATTACHED_DESIGN_URL_V2_VALUE_SCHEMA } from './schemas';
 
-import { NotFoundOperationError, OperationError } from '../../common/errors';
+import { NotFoundOperationError } from '../../common/errors';
 import { ensureString } from '../../common/string-utils';
 import type {
 	AtlassianDesign,
@@ -323,58 +324,6 @@ class JiraService {
 			);
 		}
 	};
-}
-
-export class SubmitDesignJiraOperationError extends OperationError {
-	designId?: string;
-	rejectionErrors?: { readonly message: string }[];
-	unknownIssueKeys?: string[];
-	unknownAssociations?: AtlassianAssociation[];
-
-	private constructor({
-		message,
-		designId,
-		rejectionErrors,
-		unknownIssueKeys,
-		unknownAssociations,
-	}: {
-		message: string;
-		designId?: string;
-		rejectionErrors?: { readonly message: string }[];
-		unknownIssueKeys?: string[];
-		unknownAssociations?: AtlassianAssociation[];
-	}) {
-		super(message);
-		this.designId = designId;
-		this.rejectionErrors = rejectionErrors;
-		this.unknownIssueKeys = unknownIssueKeys;
-		this.unknownAssociations = unknownAssociations;
-	}
-
-	static designRejected(
-		designId: string,
-		rejectionErrors: { readonly message: string }[],
-	): SubmitDesignJiraOperationError {
-		return new SubmitDesignJiraOperationError({
-			message: 'The design submission has been rejected',
-			designId,
-			rejectionErrors,
-		});
-	}
-
-	static unknownIssueKeys(unknownIssueKeys: string[]) {
-		return new SubmitDesignJiraOperationError({
-			message: 'The design has unknown issue keys',
-			unknownIssueKeys,
-		});
-	}
-
-	static unknownAssociations(unknownAssociations: AtlassianAssociation[]) {
-		return new SubmitDesignJiraOperationError({
-			message: 'The design has unknown associations',
-			unknownAssociations,
-		});
-	}
 }
 
 export const jiraService = new JiraService();
